requirements.txt
tensorflow-cpu==2.5.0
keras==2.4.3
imutils==0.5.4
numpy>=1.19.5
opencv-python==4.5.1.*
matplotlib==3.4.1
scipy==1.6.2
scikit-learn==0.24.1
pillow==8.2.0
streamlit==0.79.0
flask
flask-cors